backend/requirements.txt
django==4.2.11
django-ninja==1.3.0
django-cors-headers==3.13.0
django-redis==5.2.0
pyjwt==2.10.0
seldom==3.10.0